UPDATE : mysqld user mysql / add 'if' entrypoint.sh / change mariaDB query

+++ srcs/requirements/mariadb/entrypoint.sh
@@ -1,10 +1,13 @@
 #!/bin/bash
 
-cp -rp data/mysql.cnf etc/mysql/conf.d/
-cp -rp mariadb.cnf etc/mysql
-cp -rp my.cnf etc/mysql/
-service mysql start
-mysql < db.sql
-service mysql stop
-mkdir var/lib/mysql/wordpress
+FILE=/var/lib/mysql/check.txt
+if [ ! -e "$FILE" ]; then
+	touch /var/lib/mysql/check.txt
+#	mkdir var/lib/mysql/wordpress
+#	cp -rp data/mysql.cnf etc/mysql/conf.d/
+	cp -rp mariadb.cnf etc/mysql
+	cp -rp my.cnf etc/mysql/
+	service mysql start
+	mysql < db.sql
+	service mysql stop
 mysqld
